Moved repository into a separate package
