+++ server/core/dynamodb/repository/utils.go
@@ -1,4 +1,4 @@
-package dynamodb
+package repository
 
 import (
 	"context"
